Main

Documentation

main :: IO ()

test :: FilePath -> IO [Maybe String]

Extracts and tests all the examples in a file

f - filepath

Returns Returns a list of Nothing on success otherwise a descriptive string with the error.

makeAndTest :: FilePath -> IO [Maybe String]

Compiles (running ghc --make), extracts and tests all the examples in a file

f - filepath

Returns Returns a list of Nothing on success otherwise a descriptive string with the error.

tryTestCase :: FilePath -> (String, String) -> IO (Maybe String)

Tries to run the test case

f - filepath to the module in which the example will run (example, expected_result) - tuple with the string to be run and its expected value

Returns Nothing if correct otherwise it returns a descriptive string with the error.